+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Changed
+- redesigned the admin settings sections as tabbed navigation
+- changed common and security admin settings to save automatically
+- added supported languages to the Help Center link
+
 ## 10.1.2
 
 ## Changed